nord-rf-driver/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0.0
